refactor(api): MCPTransport.Port to uint16 (structural 1-65535; drop range validator)

+++ internal/cli/declarative/init.go
@@ -819,7 +819,7 @@ func writeDeclarativeMCPYAML(projectDir, name, image, description string, port i
 					Identifier:   image,
 					Transport: v1alpha1.MCPTransport{
 						Type: "http",
-						Port: uint32(port),
+						Port: uint16(port),
 						Path: "/mcp",
 					},
 				},
